Safari-609.1.11.1/LayoutTests/media/modern-media-controls/play-pause-button/play-pause-button-expected.txt

Testing the PlayPauseButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS playButton.element.localName is "button"
PASS playButton.element.classList.contains("play-pause") is true
PASS playButton.iconName is Icons.Play
PASS playButton.playing is false
PASS pauseButton.playing is true
PASS pauseButton.iconName is Icons.Pause
PASS playButton.image.element.style.webkitMaskImage.includes("macOS/Play.svg") became true
PASS pauseButton.image.element.style.webkitMaskImage.includes("macOS/Pause.svg") became true
PASS iOSPlayButton.image.element.style.webkitMaskImage.includes("iOS/Play.svg") became true
PASS iOSPauseButton.image.element.style.webkitMaskImage.includes("iOS/Pause.svg") became true

PASS successfullyParsed is true

TEST COMPLETE
